+++ spec/spec_helper.rb
@@ -3,8 +3,8 @@ $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 ENV['RAILS_ENV'] ||= 'test'
 
 require 'rails'
-require 'active_record/errors'
 require 'active_model'
+require 'active_record/errors'
 require 'action_view'
 require 'action_controller'
 require 'dummy/config/environment'
@@ -16,9 +16,3 @@ RSpec.configure do |c|
   c.infer_base_class_for_anonymous_controllers = false
   c.order = 'random'
 end
-
-class Raven
-  def capture_exception(error); end
-end
-
-class Airbrake; end